add projectProfileCard demo

+++ Spring-Security/src/main/java/com/example/springsecurityexample/match/dto/MatchRequestDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MatchRequestDto {
-    private int uid1;
+    private Long uid1;
 }
